+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## 2019-11-07
+### Added
+* Query capabilities for conditions table.
 ## 2019-10-24
 ### Added
 * Added boto3 and botocore to setup.py